Feature/47 decouple render viewer (#52)
* Views code refactoring

* Fixed runtime issues after refactoring

* Implemented rendering abstraction

* more graphic factory related changes

* Completed model view renderer refactoring

* Improved GeometryFactory instance
