+++ src/index.js
@@ -1,3 +1,6 @@
+import { GeometryFactory } from './view/util/geometryFactory'
+
+new GeometryFactory(); //instantiate singleton helper
 export {WebGLSceneModule} from './components/webGLScene.js';
